Combine <X>RowContainer and <X>Row components to remove unnecessary split
